Pulling changes in from private project
A bunch of changes were made in the private project this is based on and this commit brings this up to date with the private site. Will add some documentation that will describe functionality.

Essentially now there is a page to manage access codes, an access log review page, a settings page, and admin user login and registration.